fix: update document logic with fixing defoul adresses
