+++ src/hooks/useRegistrationForm.ts
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import { FormData } from '../types/form';
 
-
 export default function useRegistrationForm() {
   const [formData, setFormData] = useState<FormData>({
     username: '',
@@ -52,16 +51,19 @@ export default function useRegistrationForm() {
     setUseSameAddress(isChecked);
 
     if (isChecked) {
-      setFormData((prev) => ({
-        ...prev,
-        billingAddress: { ...prev.shippingAddress },
-      }));
-    } else {
-      setFormData((prev) => ({
-        ...prev,
-        billingAddress: { streetName: '', city: '', postalCode: '', country: '' },
-      }));
-    }
+  setFormData((prev) => ({
+    ...prev,
+    billingAddress: {
+      ...prev.shippingAddress,
+      defaultBillingAddress: prev.billingAddress.defaultBillingAddress ?? prev.shippingAddress.defaultShippingAddress,
+    },
+  }));
+} else {
+  setFormData((prev) => ({
+    ...prev,
+    billingAddress: { streetName: '', city: '', postalCode: '', country: '', defaultBillingAddress: false },
+  }));
+}
   };
 
   const handleDefaultShippingChange = (e: React.ChangeEvent<HTMLInputElement>) => {
